pyproject.toml
[build-system]
requires = ["setuptools>=65.0", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "voice-agent-mvp"
version = "0.1.0"
description = "Minimal voice AI agent prototype using LiveKit"
readme = "README.md"
requires-python = ">=3.11"
license = {text = "MIT"}
authors = [
    {name = "Voice Agent Team"}
]

dependencies = [
    "livekit-agents>=1.3.0,<2.0.0",
    "livekit-plugins-openai>=1.3.0,<2.0.0",
    "livekit-plugins-deepgram>=1.3.0,<2.0.0",
    "livekit-plugins-elevenlabs>=1.3.0,<2.0.0",
    "livekit-plugins-silero>=1.3.0,<2.0.0",
    "pydantic>=2.0.0,<3.0.0",
    "pydantic-settings>=2.0.0,<3.0.0",
    "httpx>=0.27.0,<2.0.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=8.0.0",
    "pytest-asyncio>=0.24.0",
    "pytest-cov>=4.1.0",
    "hypothesis>=6.100.0",
    "ruff>=0.5.0",
    "mypy>=1.10.0",
    "python-dotenv>=1.0.0",
]

[project.scripts]
voice-agent = "agent.main:main"

[tool.setuptools.packages.find]
where = ["."]
include = ["agent*"]

[tool.pytest.ini_options]
asyncio_mode = "auto"
asyncio_default_fixture_loop_scope = "function"
testpaths = ["tests"]
python_files = ["test_*.py"]
python_functions = ["test_*"]
addopts = "--cov=agent --cov-report=term-missing"

[tool.coverage.run]
branch = true
source = ["agent"]

[tool.coverage.report]
exclude_lines = [
    "pragma: no cover",
    "def __repr__",
    "raise AssertionError",
    "raise NotImplementedError",
    "if __name__ == .__main__.:",
]

[tool.ruff]
line-length = 100
target-version = "py311"

[tool.ruff.lint]
select = ["E", "F", "I", "W", "UP"]

[tool.mypy]
python_version = "3.11"
strict = true
